src/main/java/Radio.java
public class Radio {
    public int currentStation;
    public int currentVolume;

    public int getCurrentStation() {
        return currentStation;
    }
    public void setCurrentStation (int newCurrentStation) {
        if (newCurrentStation > 9) {
            newCurrentStation = 0;
        }
        if (newCurrentStation < 0) {
            newCurrentStation = 9;
        }
        currentStation = newCurrentStation;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }
    public void setCurrentVolume (int newCurrentVolume) {
        if (newCurrentVolume > 100) {
            newCurrentVolume = 100;
        }
        if (newCurrentVolume < 0) {
            newCurrentVolume = 0;
        }
        currentVolume = newCurrentVolume;
    }

    public void increaseVolume() {
        if (currentVolume < 100) {
            currentVolume = currentVolume + 1;
        }
    }

    public void downVolume() {
        if (currentVolume < 100) {
            currentVolume = currentVolume - 1;
        }
    }

    public void nextStation() {
        if (currentStation <10) {
            currentStation = currentStation + 1;
        }
    }

    public void prevStation() {
        if (currentStation <10) {
            currentStation = currentStation - 1;
        }
    }
}
